refactor: update room and vote routes, remove unused duck data, and enhance vote statistics handling
