+++ server/controllers/scenario_controller.js
@@ -6,30 +6,7 @@ class ScenarioController extends BaseController {
         super(ScenarioModel); 
     }
 
-    async updateVoteStatistics(scenario_id, answer){
-        try {
-            const scenario = await Scenario.findById(scenario_id);
-        
-            if (!scenario) {
-              console.error("Scenario not found:", scenario_id);
-              return false;
-            }
-        
-            // update number of votes
-            if (answer) {
-              scenario.votes_yes += 1;
-            } else {
-              scenario.votes_no += 1;
-            }
-        
-            // save changes
-            await scenario.save();
-            return true;
-          } catch (error) {
-            console.error("Error updating vote statistics:", error);
-            return false;
-          }
-    };
+
  
 }
 export default new ScenarioController();
