+++ server/routes/room_route.js
@@ -1,5 +1,6 @@
 import express from 'express';
 import roomController from '../controllers/room_controller.js';
+import voteController from '../controllers/vote_controller.js';
 
 const router = express.Router();
 
@@ -11,9 +12,14 @@ router.put('/:id', (req, res) => roomController.update(req, res));
 router.delete('/:id', (req, res) => roomController.delete(req, res));
 
 router.post('/join/:joinCode', (req, res) => roomController.joinRoom(req, res));
+
+// Vote-related routes (but under the /rooms namespace)
+router.post('/:room_id/answers', (req, res) => voteController.submitAnswer(req, res));
+router.get('/:room_id/statistics/:scenario_id', (req, res) => voteController.getStatistics(req, res));
+
+
 router.get('/:room_id/next-scenario/:owner_id', (req, res) => roomController.nextScenario(req, res));
-router.post('/:room_id/answers', (req, res) => roomController.submitAnswer(req, res));
-router.get('/:room_id/statistics/:scenario_id', (req, res) => roomController.getStatistics(req, res));
+
