+++ server/routes/vote_route.js
@@ -8,7 +8,12 @@ router.get('/', (req, res) => voteController.getAll(req, res));
 router.get('/:id', (req, res) => voteController.getById(req, res));
 router.put('/:id', (req, res) => voteController.update(req, res));
 router.delete('/:id', (req, res) => voteController.delete(req, res));
-router.post("/rooms/:room_id/answers", voteController.submitAnswer);
+
+router.post("/rooms/:room_id/answers", (req, res) => voteController.submitAnswer(req, res));
+router.get("/rooms/:room_id/statistics/:scenario_id", (req, res) => voteController.getStatistics(req, res));
+
+
+
 
 export default router;
